Fixed incorrect document identifier for filtering documents.
