+++ src/zaakmagazijn/api/tests/test_geef_zaakdocument_lezen.py
@@ -4,13 +4,13 @@ from io import BytesIO
 from lxml import etree
 from zeep.xsd.const import Nil
 
-from zaakmagazijn.rgbz.choices import JaNee
-from zaakmagazijn.rgbz.tests.factory_models import (
+from ...rgbz.choices import JaNee
+from ...rgbz.tests.factory_models import (
     EnkelvoudigInformatieObjectFactory, ZaakFactory,
     ZaakInformatieObjectFactory
 )
-
-from .base import BaseSoapTests
+from ..stuf.choices import BerichtcodeChoices
+from .base import BaseSoapTests, BaseTestPlatformTests
 
 
 class DMSMockMixin:
@@ -84,12 +84,12 @@ class geefZaakdocumentLezen_EdcLv01Tests(DMSMockMixin, BaseSoapTests):
                 })
             ),
             gelijk=zkn_factory.GeefZaakdocumentLezen_EDC_vraagSelectie(
-                identificatie='123456789',
+                identificatie=edc1.informatieobjectidentificatie,
             )
         )
         self.assertEqual(len(response.antwoord.object), 1)
         self.assertEqual(len(response.antwoord.object[0].isRelevantVoor), 1)
-        self.assertEqual(response.antwoord.object[0].identificatie, '123456789')
+        self.assertEqual(response.antwoord.object[0].identificatie, edc1.informatieobjectidentificatie)
         self.assertEqual(response.antwoord.object[0].titel, 'doc 1')
 
 
@@ -149,7 +149,7 @@ class geefZaakdocumentLezen_EdcLa01Tests(DMSMockMixin, BaseSoapTests):
                     })
                 ),
                 gelijk=zkn_factory.GeefZaakdocumentLezen_EDC_vraagSelectie(
-                    identificatie=self.document.identificatie,
+                    identificatie=self.document.informatieobjectidentificatie,
                 )
             )
 
@@ -299,3 +299,141 @@ class geefZaakdocumentLezen_EdcLa01Tests(DMSMockMixin, BaseSoapTests):
             1,
             namespaces=self.nsmap
         )
+
+
+class STPgeefZaakdocumentlezen_EdcLv01Tests(DMSMockMixin, BaseTestPlatformTests):
+    """
+    Precondities:
+    Scenario's voegZaakdocumentToe (P), maakZaakdocument (P) en updateZaakdocument (P) zijn succesvol uitgevoerd.
+
+    Dit scenario bevraagt documenten die in scenario's voegZaakdocumentToe (P) en maakZaakdocument (P)
+    zijn toegevoegd en (deels) in scenario updateZaakdocument (P) zijn gewijzigd.
+    """
+    porttype = 'Beantwoordvraag'
+    maxDiff = None
+    test_files_subfolder = 'stp_geefZaakdocumentlezen'
+
+    def setUp(self):
+        super().setUp()
+
+        self._dms_client.geef_inhoud.return_value = ('doc 1', BytesIO())
+
+        self.document = EnkelvoudigInformatieObjectFactory.create()
+        self.zaak = ZaakFactory.create(status_set__indicatie_laatst_gezette_status=JaNee.ja)
+        ZaakInformatieObjectFactory.create(zaak=self.zaak, informatieobject=self.document)
+
+    def _test_response(self, response):
+        self.assertEquals(response.status_code, 200, response.content)
+
+        response_root = etree.fromstring(response.content)
+        response_berichtcode = response_root.xpath(
+            '//zkn:stuurgegevens/stuf:berichtcode',
+            namespaces=self.nsmap
+        )[0].text
+        self.assertEqual(response_berichtcode, BerichtcodeChoices.la01, response.content)
+
+        response_object_element = response_root.xpath('//zkn:antwoord/zkn:object', namespaces=self.nsmap)[0]
+        response_edc_identificatie = response_object_element.xpath('zkn:identificatie', namespaces=self.nsmap)[0].text
+
+        self.assertEqual(response_edc_identificatie, self.document.informatieobjectidentificatie)
+
+    def test_geefLijstZaakdocumentenlezen_EdcLv01_01(self):
+        vraag = 'geefZaakdocumentLezen_EdcLv01_01.xml'
+        context = {
+            'voegzaakdocumenttoe_identificatie_1': self.document.informatieobjectidentificatie,
+        }
+        response = self._do_request(self.porttype, vraag, context)
+
+        self._test_response(response)
+
+    def test_geefLijstZaakdocumentenlezen_EdcLv01_03(self):
+        vraag = 'geefZaakdocumentLezen_EdcLv01_03.xml'
+        context = {
+            'voegzaakdocumenttoe_identificatie_1': self.document.informatieobjectidentificatie,
+        }
+        response = self._do_request(self.porttype, vraag, context)
+
+        self._test_response(response)
+
+    def test_geefLijstZaakdocumentenlezen_EdcLv01_05(self):
+        vraag = 'geefZaakdocumentLezen_EdcLv01_05.xml'
+        context = {
+            'voegzaakdocumenttoe_identificatie_1': self.document.informatieobjectidentificatie,
+        }
+        response = self._do_request(self.porttype, vraag, context)
+
+        self._test_response(response)
+
+    def test_geefLijstZaakdocumentenlezen_EdcLv01_07(self):
+        vraag = 'geefZaakdocumentLezen_EdcLv01_07.xml'
+        context = {
+            'voegzaakdocumenttoe_identificatie_1': self.document.informatieobjectidentificatie,
+        }
+        response = self._do_request(self.porttype, vraag, context)
+
+        self._test_response(response)
+
+    def test_geefLijstZaakdocumentenlezen_EdcLv01_09(self):
+        vraag = 'geefZaakdocumentLezen_EdcLv01_09.xml'
+        context = {
+            'voegzaakdocumenttoe_identificatie_3': self.document.informatieobjectidentificatie,
+        }
+        response = self._do_request(self.porttype, vraag, context)
+
+        self._test_response(response)
+
+    def test_geefLijstZaakdocumentenlezen_EdcLv01_11(self):
+        vraag = 'geefZaakdocumentLezen_EdcLv01_11.xml'
+        context = {
+            'maakzaakdocument_identificatie_1': self.document.informatieobjectidentificatie,
+        }
+        response = self._do_request(self.porttype, vraag, context)
+
+        self._test_response(response)
+
+    def test_geefLijstZaakdocumentenlezen_EdcLv01_13(self):
+        vraag = 'geefZaakdocumentLezen_EdcLv01_13.xml'
+        context = {
+            'voegzaakdocumenttoe_identificatie_5': self.document.informatieobjectidentificatie,
+        }
+
+        response = self._do_request(self.porttype, vraag, context)
+
+        self._test_response(response)
+
+
+class geefZaakdocumentLezen_EdcLv01RegressionTests(DMSMockMixin, BaseTestPlatformTests):
+    test_files_subfolder = 'maykin_geefZaakdocumentLezen'
+    porttype = 'Beantwoordvraag'
+
+    def setUp(self):
+        super().setUp()
+
+        self._dms_client.geef_inhoud.return_value = ('doc 1', BytesIO())
+
+    def test_empty_result_error(self):
+        """
+        See: https://taiga.maykinmedia.nl/project/haarlem-zaakmagazijn/issue/378
+        """
+        document = EnkelvoudigInformatieObjectFactory.create(
+            informatieobjectidentificatie='03923185840f-7268-44a6-95ba-085c385b0544'
+        )
+        zaak = ZaakFactory.create(status_set__indicatie_laatst_gezette_status=JaNee.ja)
+        ZaakInformatieObjectFactory.create(zaak=zaak, informatieobject=document)
+
+        vraag = 'geefZaakdocumentLezen_EdcLv01_taiga378.xml'
+        response = self._do_request(self.porttype, vraag)
+
+        self.assertEquals(response.status_code, 200, response.content)
+
+        response_root = etree.fromstring(response.content)
+        response_berichtcode = response_root.xpath(
+            '//zkn:stuurgegevens/stuf:berichtcode',
+            namespaces=self.nsmap
+        )[0].text
+        self.assertEqual(response_berichtcode, BerichtcodeChoices.la01, response.content)
+
+        response_object_element = response_root.xpath('//zkn:antwoord/zkn:object', namespaces=self.nsmap)[0]
+        response_edc_identificatie = response_object_element.xpath('zkn:identificatie', namespaces=self.nsmap)[0].text
+
+        self.assertEqual(response_edc_identificatie, document.informatieobjectidentificatie)
